+++ projectile.el
@@ -1337,6 +1337,12 @@ With a prefix argument ARG prompts you for a directory on which the search is pe
   (projectile-with-default-dir (projectile-project-root)
     (call-interactively 'shell-command)))
 
+(defun projectile-run-async-shell-command-in-root ()
+  "Invoke `async-shell-command' in the project's root."
+  (interactive)
+  (projectile-with-default-dir (projectile-project-root)
+    (call-interactively 'async-shell-command)))
+
 (defun projectile-files-in-project-directory (directory)
   "Return a list of files in DIRECTORY."
   (let ((dir (file-relative-name (expand-file-name directory) (projectile-project-root))))
@@ -1809,6 +1815,7 @@ is chosen."
       (define-key prefix-map (kbd "4 f") 'projectile-find-file-other-window)
       (define-key prefix-map (kbd "4 t") 'projectile-find-implementation-or-test-other-window)
       (define-key prefix-map (kbd "!") 'projectile-run-shell-command-in-root)
+      (define-key prefix-map (kbd "&") 'projectile-run-async-shell-command-in-root)
       (define-key prefix-map (kbd "a") 'projectile-ack)
       (define-key prefix-map (kbd "A") 'projectile-ag)
       (define-key prefix-map (kbd "b") 'projectile-switch-to-buffer)